backend/services/evidence-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0

